refactor: organize app shell and shared theme
