+++ frontend/src/app/App.tsx
@@ -1,7 +1,7 @@
-import { RootProvider } from "@/context/RootProvider";
-import { router } from "@/router";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { RouterProvider } from "react-router";
+import { RootProvider } from "./RootProvider";
+import { router } from "./router";
 
 export default function App() {
   return (
