+++ frontend/src/app/RootProvider.tsx
@@ -1,9 +1,9 @@
 import { type ReactNode } from "react";
-import { ThemeProvider } from "./ThemeContext";
 import { AuthProvider } from "@/features/auth";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "@/lib/query/queryClient";
 import { TooltipProvider } from "@/components/ui/tooltip";
+import { ThemeProvider } from "@/shared/theme/ThemeContext";
 
 export const RootProvider = ({ children }: { children: ReactNode }) => {
   return (
