+++ frontend/src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import { PostHogProvider } from "posthog-js/react";
-import { POSTHOG_KEY, POSTHOG_URL, MODE } from "./constants";
+import { POSTHOG_KEY, POSTHOG_URL, MODE } from "./config";
 import "@/lib/logger"; // Initialize logger
-import App from "./App";
+import App from "./app/App";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
